Make tensorflow models thread-safe in Flask
On running a prediction from a flask resource, tensorflow gives the error that "<a tensor> is not an element of this graph". The reason is that TF is not thread-safe. We apply the workaround to store the default graph to a variable when loading the model, and making it the default graph explicitly on prediction.